Rename the history management to archive management, add modify archive popup location
